pyproject.toml
[build-system]
# Minimum requirements for the build system to execute.
requires = ["setuptools==75.8.2", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yt_navigator"
dynamic = ["version"]
description = "A project powered by AI to help you find information on specific YouTube channels."
readme = "README.md"
requires-python = ">=3.13.2"
license = { text = "MIT" }
dependencies = [
    "django==5.1.7",
    "gunicorn==21.2.0",
    "psycopg2-binary==2.9.10",
    "asyncpg==0.30.0",
    "openai>=1.54.0",
    "yt-dlp>=2024.12.6",
    "rank_bm25==0.2.2",
    "python-dotenv==1.0.1",
    "structlog==25.1.0",
    "youtube-transcript-api==0.6.3",
    "scrapetube==2.5.1",
    "bs4==0.0.2",
    "httpx==0.28.1",
    "langchain-postgres",
    "langchain-community==0.3.17",
    "langgraph==0.3.5",
    "langchain-openai>=0.2.0",
    "langgraph-checkpoint-postgres==2.0.16",
    "markdown2==2.5.3",
    "scikit-learn>=1.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-django",
    "black",
    "isort",
    "flake8",
    "ruff",
    "djlint==1.36.4",
]

[tool.setuptools]
packages = ["yt_navigator", "app"]

[tool.setuptools.dynamic]
version = { file = "version.txt" }


[tool.pytest.ini_options]
markers = ["slow: marks tests as slow (deselect with '-m \"not slow\"')"]
DJANGO_SETTINGS_MODULE = "yt_navigator.settings"
python_files = ["test_*.py", "*_test.py", "tests.py"]

[tool.black]
line-length = 119
exclude = "venv|migrations"

[tool.flake8]
docstring-convention = "all"
ignore = ["D107", "D212", "E501", "W503", "W605", "D203", "D100"]
exclude = "venv|migrations"
max-line-length = 119

# radon
radon-max-cc = 10

[tool.isort]
profile = "black"
multi_line_output = "VERTICAL_HANGING_INDENT"
force_grid_wrap = 2
line_length = 119
skip = ["migrations", "venv"]

[tool.pylint."messages control"]
disable = [
    "line-too-long",
    "trailing-whitespace",
    "missing-function-docstring",
    "consider-using-f-string",
    "import-error",
    "too-few-public-methods",
    "redefined-outer-name",
]

[tool.pylint.master]
ignore = "migrations"

[tool.ruff]
line-length = 119
exclude = ["migrations", "*.ipynb", "venv"]

[tool.ruff.lint]
# Enable flake8-bugbear (`B`) rules and docstring (`D`) rules
select = ["E", "F", "B", "ERA", "D"]
# Never enforce `E501` (line length violations).
ignore = ["E501", "F401", "D203", "D213"]
# Avoid trying to fix flake8-bugbear (`B`) violations.
unfixable = ["B"]

[tool.ruff.lint.pydocstyle]
convention = "google"

# Ignore `E402` (import violations) in all `__init__.py` files
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402"]
